+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 ### Logins
 - fix `count_by_origin` and `count_by_form_action_origin` with punicode origins
 
+### Places
+- `places::storage::history_metadata::get_most_recent_search_entries()` was added to fetch the most recent search entries in history metadata. ([#7104](https://github.com/mozilla/application-services/pull/7104))
+
 # v146.0 (_2025-11-10_)
 
 ## ✨ What's New ✨